refactor: move shard layout API to EpochManagerAdapter (#7660)
Part of https://github.com/near/nearcore/issues/6910, depends on https://github.com/near/nearcore/pull/7657

Must be reviewed per commit.
